data-preprocess-server/requirements.txt
flask==2.3.2
pandas==1.5.2
pymongo==4.4.1
requests==2.31.0
python-dotenv==1.0.0
flask-restful==0.3.10
flask-swagger-ui
apiflask==2.1.0